Finishes display widget; renames admin widget
