+++ templates/display-widget.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * The template for the hours display widget.
+ *
+ * @package MITlib Pull Hours
+ * @since 0.2.0
+ */
+
+?>
+
+<p class="hours-today">Today's hours:
+	<span 
+		style="display:inline-block; margin-bottom: 0;"
+		data-location-hours="<?php echo esc_attr( $instance['location_slug'] ); ?>"></span><br />
+	<a href="/hours">See all hours</a>
+</p>
